requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
mysqlclient==2.2.4
python-dotenv==1.0.1
pytest==8.3.4
sqlalchemy_utils==0.41.2
boto3==1.34.58
watchtower==3.4.0
logger==1.4
statsd==4.0.1